refactor : test코드 model 패키지 생성
