+++ src/test/java/coordinate/model/PointTest.java
@@ -1,0 +1,5 @@
+package coordinate.model;
+
+public class PointTest {
+
+}
